Move to jest and azure.
